macOS: support reloading temporary config for testing
